+++ Packages/com.seahyx.xruitoolkit/Runtime/Core/Visual Effects/Visual Effect Runtimes/RectPositionVisualEffect.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace XRUIToolkit.Core.VisualEffect
 {
-	[CreateAssetMenu(fileName = "RectPositionEffect", menuName = "HHUI Toolkit/Visual Effects/RectTransform Position Effect", order = 2)]
+	[CreateAssetMenu(fileName = "RectPositionVFX", menuName = "HHUI Toolkit/Visual Effects/RectTransform Position VFX", order = 2)]
 	public class RectPositionVisualEffect : BaseAnimatedVisualEffect
 	{
 		protected enum TransformMode