+++ Packages/com.seahyx.xruitoolkit/Runtime/Core/Visual Effects/Visual Effect Runtimes/TextVisualEffect.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace XRUIToolkit.Core.VisualEffect
 {
-	[CreateAssetMenu(fileName = "TextVisualEffect", menuName = "HHUI Toolkit/Visual Effects/Text Visual Effect", order = 0)]
+	[CreateAssetMenu(fileName = "TextVFX", menuName = "HHUI Toolkit/Visual Effects/Text VFX", order = 0)]
 	public class TextVisualEffect : BaseVisualEffect
 	{
 		[SerializeField,